Upadte wpulse.cpp/v86boot.cpp/86Duino IDE

+++ hardware/86duino/x86/cores/arduino/stdint.h
@@ -3,9 +3,9 @@ typedef signed char int8_t;
 
 typedef unsigned char uint8_t;
 
-typedef signed int int16_t;
+typedef signed short int16_t;
 
-typedef unsigned int uint16_t;
+typedef unsigned short uint16_t;
 
 typedef signed long int int32_t;
